requirements.txt
backoff==1.6.0
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
dill==0.2.8.2
enum==0.4.6
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JSONRPC==0.3.1
funcsigs==1.0.2
gevent==1.3.5
geventhttpclient-wheels==1.3.1.dev1
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
locustio==0.9.0
MarkupSafe==1.0
matplotlib==2.2.3
mock==2.0.0
msgpack-python==0.5.6
multiprocess==0.70.6.1
numpy==1.14.5
pandas==0.23.3
pbr==4.1.0
pox==0.2.4
ppft==1.6.4.8
psutil==5.4.6
psycopg2-binary==2.7.5
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.0
requests==2.19.1
setproctitle==1.1.10
six==1.11.0
subprocess32==3.5.2
urllib3==1.23
Werkzeug==0.14.1
